Staff-Leave-Management 4/flask-backend-sqlite/requirements.txt
Flask==3.1.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
